backend/requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.11
bidict==0.23.1
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-SocketIO==5.3.6
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-engineio==4.9.0
python-socketio==5.11.2
RPi.GPIO==0.7.1
RPLCD==1.3.1
simple-websocket==1.0.0
smbus2==0.4.3
spidev==3.6
Werkzeug==3.0.2
wsproto==1.2.0
